+++ packages/nimble-components/src/patterns/button/styles.ts
@@ -103,14 +103,27 @@ export const styles = css`
     }
 
     :host([content-hidden]) .content {
-        display: none;
+        ${
+            /**
+             * Hide content visually while keeping it screen reader-accessible.
+             * Source: https://webaim.org/techniques/css/invisiblecontent/#techniques
+             * See discussion here: https://github.com/microsoft/fast/issues/5740#issuecomment-1068195035
+             */
+            ''
+        }
+        display: inline-block;
+        height: 1px;
+        width: 1px;
+        position: absolute;
+        margin: -1px;
+        clip: rect(1px, 1px, 1px, 1px);
+        clip-path: inset(50%);
+        overflow: hidden;
+        padding: 0;
     }
 
     [part='start'] {
         display: contents;
-    }
-
-    slot[name='start']::slotted(*) {
         ${iconColor.cssCustomProperty}: ${buttonLabelFontColor};
     }
 
@@ -119,6 +132,11 @@ export const styles = css`
     }
 
     [part='end'] {
+        display: contents;
+        ${iconColor.cssCustomProperty}: ${buttonLabelFontColor};
+    }
+
+    :host([content-hidden]) [part='end'] {
         display: none;
     }
 `
@@ -191,13 +209,17 @@ export const styles = css`
                 }
 
                 .control:hover {
-                    background-color: rgba(${borderRgbPartialColor}, 0.1);
+                    background-color: transparent;
                     border-color: ${borderHoverColor};
                 }
 
                 .control${focusVisible} {
                     background-color: rgba(${borderRgbPartialColor}, 0.1);
                     border-color: ${borderHoverColor};
+                }
+
+                .control${focusVisible}:hover {
+                    background-color: transparent;
                 }
 
                 .control:active {